Se se modifico creacion de cuentas de usuario

+++ database/migrations/2022_06_19_002844_crear_cuentas_bancarias.php
@@ -17,7 +17,7 @@ class CrearCuentasBancarias extends Migration
             $table->id();
             $table->string('name');
             $table->string('email')->unique();
-            $table->integer('cuenta');
+            $table->bigInteger('cuenta');
             $table->decimal('saldo', 10, 2);
             $table->foreignId('user_id')->constrained('users')->onUpdate('cascade')->onDelete('cascade');
             $table->timestamps();
